+++ CHANGELOG.md
@@ -14,12 +14,14 @@ You can also check the
 - Features
 
   - Added stacked horizontal charts for power stability to the sunshine charts
+  - Added Error States & enhanced existing hint styles
 
 - Storybook
 
   - Added stacked horizontal chart to storybook
   - SAIDI / SAIFI indicators can be shown on the map (behind "sunshine" flag)
-  - 
+  - Added Error States to storybook
+
 - Fixes
 
   - Changes to Docker image to be able to use duckdb (alpine based to debian based)