+++ CHANGELOG.md
@@ -11,9 +11,22 @@ You can also check the
 
 # Unreleased
 
+- Features
+
+  - Added stacked horizontal charts for power stability to the sunshine charts
+
+- Storybook
+
+  - Added stacked horizontal chart to storybook
+  - SAIDI / SAIFI indicators can be shown on the map (behind "sunshine" flag)
+  - 
 - Fixes
+
   - Changes to Docker image to be able to use duckdb (alpine based to debian based)
   - Correct fetching of power stability metrics through graphql queries
+
+- Maintenance
+  - Updated tpo-deepl see README.md for details `https://www.npmjs.com/package/tpo-deepl`
 
 # 2.0.0 - 2025-06-11
 
@@ -26,6 +39,7 @@ You can also check the
   - Load data through urql in sunshine pages, urql client cache is primed by the server
   - Added Scatterplot Chart to sunshine charts
   - Ability to change network level on costs and tariffs > network costs page.
+
 - Fixes
   - Fixed map layout issues on mobile
   - Dropdown adjustments
@@ -39,6 +53,7 @@ You can also check the
   - Add new Scatterplot Chart
 - Styles
   - Update Design System to match the Bund new CI system
+  - Improved map animations / transitions
 - Maintenance
   - Added Sentry
   - Split Static checks and e2e tests